# swagger.yaml
openapi: 3.0.3
info:
  title: Word Association API
  description: |
    Word2Vecベクトルモデルを使用して、入力されたキーワードから世代数に応じて連想される言葉を返すAPIです。
    
    ## 世代数システム
    - **世代数2**: 入力キーワードから6つの連想語を取得
    - **世代数3以降**: 前世代の各単語から3つずつ連想語を階層的に取得
    - 結果は世代構造として整理されて返されます
    
    ## モデル情報
    - **語彙数**: 1,015,474語
    - **ベクトル次元**: 200次元
    - **モデルファイル**: entity_vector/entity_vector.model.txt
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server

paths:
  /api/v1/associate:
    post:
      summary: 連想語取得
      description: |
        指定されたキーワードから世代数に応じて連想される言葉を取得します
        
        **世代数の仕組み:**
        - **世代数2**: 入力キーワードから連想語6つを取得
        - **世代数3以降**: 前世代の各単語から3つずつ連想語を取得
        
        **例:**
        - 世代数2 → 6個の連想語
        - 世代数3 → 6個 + (6×3) = 24個の連想語  
        - 世代数4 → 6個 + (6×3) + (18×3) = 78個の連想語
      operationId: getAssociatedWords
      tags:
        - Association
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssociationRequest'
            examples:
              generation_2:
                summary: 世代数2の例
                value:
                  keyword: "犬"
                  generation: 2
                  threshold: 0.5
              generation_3:
                summary: 世代数3の例
                value:
                  keyword: "料理"
                  generation: 3
                  threshold: 0.7
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssociationResponse'
              examples:
                generation_2_response:
                  summary: 世代数2の成功例
                  value:
                    status: "success"
                    keyword: "犬"
                    generation: 2
                    generations:
                      - generation_number: 2
                        parent_word: "犬"
                        results:
                          - word: "猫"
                            similarity: 0.89
                          - word: "動物"
                            similarity: 0.85
                          - word: "ペット"
                            similarity: 0.82
                          - word: "散歩"
                            similarity: 0.78
                          - word: "飼い主"
                            similarity: 0.75
                          - word: "しっぽ"
                            similarity: 0.72
                        count: 6
                    total_count: 6
                generation_3_response:
                  summary: 世代数3の成功例
                  value:
                    status: "success"
                    keyword: "犬"
                    generation: 3
                    generations:
                      - generation_number: 2
                        parent_word: "犬"
                        results:
                          - word: "猫"
                            similarity: 0.89
                          - word: "動物"
                            similarity: 0.85
                          - word: "ペット"
                            similarity: 0.82
                          - word: "散歩"
                            similarity: 0.78
                          - word: "飼い主"
                            similarity: 0.75
                          - word: "しっぽ"
                            similarity: 0.72
                        count: 6
                      - generation_number: 3
                        parent_word: "猫"
                        results:
                          - word: "ネコ"
                            similarity: 0.95
                          - word: "魚"
                            similarity: 0.78
                          - word: "毛玉"
                            similarity: 0.65
                        count: 3
                      - generation_number: 3
                        parent_word: "動物"
                        results:
                          - word: "生き物"
                            similarity: 0.92
                          - word: "野生"
                            similarity: 0.83
                          - word: "哺乳類"
                            similarity: 0.79
                        count: 3
                      - generation_number: 3
                        parent_word: "ペット"
                        results:
                          - word: "飼育"
                            similarity: 0.88
                          - word: "家族"
                            similarity: 0.76
                          - word: "癒し"
                            similarity: 0.71
                        count: 3
                    total_count: 18
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_generation:
                  summary: 世代数無効
                  value:
                    status: "error"
                    error_code: "INVALID_PARAMETER"
                    message: "世代数は2以上5以下で指定してください"
                invalid_threshold:
                  summary: 閾値無効
                  value:
                    status: "error"
                    error_code: "INVALID_PARAMETER"
                    message: "thresholdは0.0から1.0の間で指定してください"
                keyword_required:
                  summary: キーワード必須
                  value:
                    status: "error"
                    error_code: "KEYWORD_REQUIRED"
                    message: "keywordパラメータは必須です"
        '404':
          description: キーワードが見つからない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                not_found:
                  summary: キーワード未発見
                  value:
                    status: "error"
                    error_code: "KEYWORD_NOT_FOUND"
                    message: "指定されたキーワードがモデルに存在しません"
        '429':
          description: レート制限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rate_limit:
                  summary: レート制限
                  value:
                    status: "error"
                    error_code: "RATE_LIMIT_EXCEEDED"
                    message: "リクエスト制限を超えました。しばらく待ってから再試行してください"
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internal_error:
                  summary: 内部エラー
                  value:
                    status: "error"
                    error_code: "INTERNAL_ERROR"
                    message: "サーバー内部エラーが発生しました"
        '503':
          description: サービス利用不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                model_error:
                  summary: モデル読み込みエラー
                  value:
                    status: "error"
                    error_code: "MODEL_LOAD_ERROR"
                    message: "モデルの読み込みに失敗しました"

  /api/v1/model/info:
    get:
      summary: モデル情報取得
      description: 使用中のWord2Vecモデルの情報を取得します
      operationId: getModelInfo
      tags:
        - Model
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelInfoResponse'
              examples:
                success:
                  summary: モデル情報
                  value:
                    status: "success"
                    model_info:
                      vocabulary_size: 1015474
                      vector_dimension: 200
                      model_type: "word2vec"
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


components:
  schemas:
    AssociationRequest:
      type: object
      required:
        - keyword
        - generation
      properties:
        keyword:
          type: string
          description: 連想語を取得したいキーワード
          minLength: 1
          maxLength: 100
          example: "犬"
        generation:
          type: integer
          description: |
            世代数（2以上）
            - 世代数2: 入力キーワードから連想語6つを取得
            - 世代数3以降: 前世代の各単語から3つずつ連想語を取得
          minimum: 2
          maximum: 5
          example: 2
        threshold:
          type: number
          format: float
          description: 類似度の閾値
          minimum: 0.0
          maximum: 1.0
          default: 0.5
          example: 0.7

    AssociationResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          description: レスポンスステータス
          example: "success"
        keyword:
          type: string
          description: 入力されたキーワード
          example: "犬"
        generation:
          type: integer
          description: 指定された世代数
          example: 2
        generations:
          type: array
          description: 世代ごとの連想語リスト
          items:
            $ref: '#/components/schemas/Generation'
        total_count:
          type: integer
          description: 全世代の結果総数
          minimum: 0
          example: 6

    Generation:
      type: object
      properties:
        generation_number:
          type: integer
          description: 世代番号
          minimum: 1
          example: 2
        parent_word:
          type: string
          description: 親となるキーワード（第1世代は入力キーワード）
          example: "犬"
        results:
          type: array
          description: この世代の連想語リスト
          items:
            $ref: '#/components/schemas/AssociationResult'
        count:
          type: integer
          description: この世代の結果数
          minimum: 0
          example: 6

    AssociationResult:
      type: object
      properties:
        word:
          type: string
          description: 連想語
          example: "猫"
        similarity:
          type: number
          format: float
          description: 類似度スコア（0.0-1.0）
          minimum: 0.0
          maximum: 1.0
          example: 0.89

    ModelInfoResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          description: レスポンスステータス
          example: "success"
        model_info:
          $ref: '#/components/schemas/ModelInfo'

    ModelInfo:
      type: object
      properties:
        vocabulary_size:
          type: integer
          description: 語彙数
          example: 1015474
        vector_dimension:
          type: integer
          description: ベクトル次元数
          example: 200
        model_type:
          type: string
          description: モデルタイプ
          example: "word2vec"


    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          description: レスポンスステータス
          example: "error"
        error_code:
          type: string
          description: エラーコード
          enum:
            - KEYWORD_NOT_FOUND
            - INVALID_PARAMETER
            - KEYWORD_REQUIRED
            - INTERNAL_ERROR
            - MODEL_LOAD_ERROR
            - RATE_LIMIT_EXCEEDED
          example: "KEYWORD_NOT_FOUND"
        message:
          type: string
          description: エラーメッセージ
          example: "指定されたキーワードがモデルに存在しません"

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API キー認証（オプション）

security:
  - {}
  - ApiKeyAuth: []

tags:
  - name: Association
    description: 連想語関連のAPI
  - name: Model
    description: モデル情報関連のAPI

externalDocs:
  description: 詳細なドキュメント
  url: https://docs.example.com/word-association-api